Removed RocketChat.ts n moved its methods to Room.ts
